requirements.txt
accounts==0.2.4
admin==0.0.1
aioredis==1.3.1
alembic==1.7.5
amqp==5.1.0
anyio==3.4.0
api==0.0.7
appdirs==1.4.3
apturl==0.5.2
asgiref==3.4.1
astroid==2.4.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
automium==0.2.6
automium-web==0.1.1
autopep8==1.5.4
backports.entry-points-selectable==1.1.1
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.10.0
billiard==3.6.4.0
black==22.1.0
bleach==4.1.0
blinker==1.4
bootstrap4==0.1.0
boto==2.49.0
boto3==1.14.56
botocore==1.17.56
Brlapi==0.7.0
CacheControl==0.12.6
cachetools==5.0.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
channels==3.0.4
channels-redis==3.3.1
chardet==3.0.4
charset-normalizer==2.0.9
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
collection==0.1.6
colorama==0.4.4
command-not-found==0.3
constantly==15.1.0
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cssselect==1.1.0
cssutils==2.3.0
cupshelpers==1.0
daphne==3.0.2
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.2.12
django-allauth==0.48.0
django-bootstrap5==21.3
django-celery-beat==2.2.1
django-celery-results==2.3.0
django-channels-notifications==1.1
django-cors-headers==3.5.0
django-crispy-forms==1.11.1
django-email-users==1.1.0
django-filter==21.1
django-jazzmin==2.4.8
django-model-utils==4.2.0
django-notifications-hq==1.6.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-smart-selects==1.5.9
django-storages==1.10
django-summernote==0.8.20.0
django-templated-mail==1.1.1
django-timezone-field==4.2.3
django-user-management==18.0.0
djangorestframework==3.13.1
djangorestframework-jsonapi==4.3.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dnspython==2.1.0
docker==5.0.3
docutils==0.15.2
drf-yasg==1.20.0
duplicity==0.8.12.0
ecdsa==0.17.0
email-validator==1.1.3
emails==0.5.15
entrypoints==0.3
fastapi==0.70.1
fasteners==0.14.1
filelock==3.4.0
flake8==4.0.1
Flask==2.0.2
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
h11==0.9.0
hiredis==2.0.0
html5lib==1.0.1
httplib2==0.14.0
httptools==0.1.2
hyperlink==21.0.0
identify==2.4.11
idna==3.3
image==1.5.33
importlib-metadata==1.5.0
importlib-resources==5.4.0
incremental==21.3.0
incuna-mail==4.1.1
incuna-pigeon==0.1.0
inflect==5.3.0
inflection==0.5.1
iniconfig==1.1.1
ipaddr==2.2.0
isort==5.4.2
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
jsonfield==3.1.0
keyring==18.0.1
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
legacy==0.1.6
lockfile==0.12.2
louis==3.12.0
lxml==4.7.1
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.3
mypy-extensions==0.4.3
netifaces==0.10.4
nodeenv==1.6.0
nose==1.3.7
oauthlib==3.1.1
olefile==0.46
openapi-codec==1.3.2
packaging==21.3
paramiko==2.6.0
passlib==1.7.4
pathspec==0.9.0
pbr==5.8.1
pep517==0.8.2
permission==0.4.1
pexpect==4.6.0
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
pre-commit==2.17.0
premailer==3.10.0
progress==1.5
prompt-toolkit==3.0.29
protobuf==3.6.1
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.21
pycups==1.9.73
pydantic==1.9.0
pyflakes==2.4.0
PyGObject==3.36.0
PyJWT==2.3.0
pylint==2.12.2
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
pyRFC3339==1.1
pytest==6.2.5
python-apt==2.0.0+ubuntu0.20.4.6
python-crontab==2.6.0
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-decouple==3.6
python-dotenv==0.19.0
python-jose==3.3.0
python-memcached==1.59
python-multipart==0.0.5
python3-openid==3.2.0
pytoml==0.1.21
pytz==2021.3
pyxdg==0.26
PyYAML==5.3.1
raven==6.10.0
reportlab==3.5.34
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
retrying==1.3.3
rsa==4.8
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.6
s3transfer==0.3.3
SecretStorage==2.3.1
service-identity==21.1.0
simplejson==3.16.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
sqlparse==0.4.2
starlette==0.16.0
swapper==1.3.0
systemd-python==234
tenacity==6.3.1
testresources==2.0.1
toml==0.10.1
tomli==2.0.1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==4.0.1
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.7
usb-creator==0.3.7
utils==1.0.1
uvicorn==0.11.8
uvloop==0.16.0
views==0.3
vine==5.0.0
virtualenv==20.10.0
wadllib==1.3.3
waitress==2.1.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
websockets==8.1
Werkzeug==2.0.2
whitenoise==6.0.0
wrapt==1.12.1
xkit==0.0.0
zipp==3.7.0
zope.interface==5.4.0
